Support authentication with environment group identities
